refactor login component and fix styling
