+++ app/login/layout.js
@@ -11,7 +11,9 @@ export const metadata = {
 export default function RootLayout({ children }) {
 	return (
 		<html lang="en">
-			<body className={`${inter.className}`}>{children}</body>
+			<body style={{ minWidth: '375px' }} className={`${inter.className}`}>
+				{children}
+			</body>
 		</html>
 	)
 }
