--- app/login/login.module.css
@@ -1,79 +1,0 @@
-.maincard {
-    height: 35rem;
-    width: 32rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-    padding: 20px;
-    border-radius: 20px;
-    box-shadow: 0 2px 4px rgba(0,0,0, 0.1);
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    padding: 0rem 1rem;
-}
-
-.p_info{
-    margin-top: 0;
-    margin-bottom: 2rem;
-}
-
-.h1_title{
-    margin-bottom: 0.3rem;
-}
-
-.input {
-    margin-top: 0.5rem;
-    padding: 10px;
-    margin-bottom: 1rem;
-    border: none;
-    border-radius: 8px;
-    background-color: #f6f6f6;
-    color: #333;
-    font-size: 16px;
-    box-shadow: 0 2px 4px rgba(0,0,0, 0.1);
-    transition: box-shadow 0.3s ease;
-}
-
-.input:focus {
-    outline: none;
-    box-shadow: 0 4px 8px rgba(0,0,0, 0.1);
-}
-
-.btn_access {
-    padding: 0.7rem;
-    border: none;
-    border-radius: 20px;
-    background-color: #43975B;
-    color: white;
-    margin-top: 1rem;
-    margin-bottom: 1rem;
-    font-size: 12pt;
-    cursor: pointer;
-}
-
-.btn_access:hover{
-    background-color: #4FAA68;
-}
-
-.btn_comeback {
-    padding: 0.7rem;
-    border: none;
-    border-radius: 20px;
-    background-color: #9f9e9e;
-    color: white;
-    font-size: 12pt;
-    cursor: pointer;
-}
-
-.btn_comeback:hover {
-    background-color: #9f9e9ed2;
-}
-
-.input_error{
-    border: solid 1px red;
-}
